wip: cache package download

+++ fs/btrfs/box/provision.sh
@@ -3,6 +3,8 @@
 set -euo pipefail
 IFS=$'\n\t'
 
+CACHE_DIR=/vagrant/.cache
+
 apt_packages() {
     apt-get update && \
         export DEBIAN_FRONTEND=noninteractive && \
@@ -11,14 +13,18 @@ apt_packages() {
             uuid-runtime python3-pip libsqlite3-dev
 }
 
-bonnie() {
+bonnie_install() {
     BONNIE_VERSION=2.00b
-    DIR=./bonnie_install
+    DIR=$CACHE_DIR/bonnie_install
 
-    wget https://github.com/bachm44/bonnie-plus-plus/archive/refs/tags/$BONNIE_VERSION.tar.gz
-    mkdir -pv $DIR
-    tar -xf *.tar.gz -C $DIR --strip-components=1
-    rm *.tar.gz
+    if [ ! -d $DIR ]
+    then
+        wget https://github.com/bachm44/bonnie-plus-plus/archive/refs/tags/$BONNIE_VERSION.tar.gz
+        mkdir -pv $DIR
+        tar -xf *.tar.gz -C $DIR --strip-components=1
+        rm *.tar.gz
+    fi
+
     pushd $DIR
         make install
     popd
@@ -26,12 +32,16 @@ bonnie() {
 
 fio_install() {
     FIO_VERSION=fio-3.33
-    DIR=./fio_install
+    DIR=$CACHE_DIR/fio_install
 
-    wget https://github.com/axboe/fio/archive/refs/tags/$FIO_VERSION.tar.gz
-    mkdir -pv $DIR
-    tar -xf *.tar.gz -C $DIR --strip-components=1
-    rm *.tar.gz
+    if [ ! -d $DIR ]
+    then
+        wget https://github.com/axboe/fio/archive/refs/tags/$FIO_VERSION.tar.gz
+        mkdir -pv $DIR
+        tar -xf *.tar.gz -C $DIR --strip-components=1
+        rm *.tar.gz
+    fi
+
     pushd $DIR
         ./configure
         make -j2
@@ -41,12 +51,16 @@ fio_install() {
 
 gen_file_install() {
     GEN_FILE_VERSION=1.0.5-dev-aa8ab56514bd6b4bf7ac5669b620c53b604fcf82
-    DIR=./gen_file_install
+    DIR=$CACHE_DIR/gen_file_install
 
-    wget https://github.com/bachm44/gen_file/releases/download/$GEN_FILE_VERSION/gen_file-$GEN_FILE_VERSION.tar.gz
-    mkdir -pv $DIR
-    tar -xf *.tar.gz -C $DIR --strip-components=1
-    rm *.tar.gz
+    if [ ! -d $DIR ]
+    then
+        wget https://github.com/bachm44/gen_file/releases/download/$GEN_FILE_VERSION/gen_file-$GEN_FILE_VERSION.tar.gz
+        mkdir -pv $DIR
+        tar -xf *.tar.gz -C $DIR --strip-components=1
+        rm *.tar.gz
+    fi
+
     pushd $DIR
         ./configure
         make -j2
@@ -60,19 +74,43 @@ kernel_install() {
     KERNEL_HEADERS=linux-headers-6.1.0-32b0e7d60_6.1.0-l_amd64.deb
     KERNEL_LIBC=linux-libc-dev_6.1.0-l_amd64.deb
 
-    wget https://github.com/bachm44/nilfs-dedup/releases/download/$RELEASE/$KERNEL_IMAGE
-    wget https://github.com/bachm44/nilfs-dedup/releases/download/$RELEASE/$KERNEL_HEADERS
-    wget https://github.com/bachm44/nilfs-dedup/releases/download/$RELEASE/$KERNEL_LIBC
+    DIR=$CACHE_DIR/linux_install
 
-    dpkg -i $KERNEL_IMAGE
-    dpkg -i $KERNEL_HEADERS
-    dpkg -i $KERNEL_LIBC
+    if [ ! -d $DIR]
+    then
+        mkdir -pv $DIR
+    fi
+
+    if [ ! -f $DIR/$KERNEL_IMAGE ]
+    then
+        wget https://github.com/bachm44/nilfs-dedup/releases/download/$RELEASE/$KERNEL_IMAGE
+        mv -v $KERNEL_IMAGE $DIR/$KERNEL_IMAGE
+    fi
+
+    if [ ! -f $DIR/$KERNEL_HEADERS ]
+    then
+        wget https://github.com/bachm44/nilfs-dedup/releases/download/$RELEASE/$KERNEL_HEADERS
+        mv -v $KERNEL_HEADERS $DIR/$KERNEL_HEADERS
+    fi
+
+    if [ ! -f $DIR/$KERNEL_LIBC ]
+    then
+        wget https://github.com/bachm44/nilfs-dedup/releases/download/$RELEASE/$KERNEL_LIBC
+        mv -v $KERNEL_LIBC $DIR/$KERNEL_LIBC
+    fi
+
+    dpkg -i $DIR/$KERNEL_IMAGE
+    dpkg -i $DIR/$KERNEL_HEADERS
+    dpkg -i $DIR/$KERNEL_LIBC
 }
 
 bees_install() {
-    DIR=./bees
+    DIR=$CACHE_DIR/bees
 
-    git clone --branch v0.9.3 --depth 1 https://github.com/Zygo/bees.git
+    if [ ! -d $DIR ]
+    then
+        git clone --branch v0.9.3 --depth 1 https://github.com/Zygo/bees.git $DIR
+    fi
 
     pushd $DIR
         make -j2
@@ -81,9 +119,13 @@ bees_install() {
 }
 
 dduper_install() {
-    DIR=./dduper
+    DIR=$CACHE_DIR/dduper
     COMMIT=11b78558f1b1677ce9407909cecaeb3374828adb
-    git clone https://github.com/Lakshmipathi/dduper.git
+
+    if [ ! -d $DIR ]
+    then
+        git clone https://github.com/Lakshmipathi/dduper.git $DIR
+    fi
 
     pushd $DIR
         git checkout $COMMIT
@@ -102,7 +144,7 @@ fix_keys() {
 
 main() {
     apt_packages
-    bonnie
+    bonnie_install
     fio_install
     gen_file_install
     kernel_install
